+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+4.1.1 - 2017-09-05
+====================
+- added -f or --csvfileds option to add fields name to the first row of csv output file.
+
+4.1 - 2017-08-07
+====================
+- added VirusTotal support. Setting the API_KEY within the config.json file
+
 4.0 - 2017-02-03
 ====================
 - release v.4.0.0